Add settings to mute sound
Add settings to mute sound. Upgrade data model.
Use new settings in controller.

Background: it is wished to mute sound but
keeping the url streams intact. A new button on the
front page should be made available.
It should be possible to hide the current button, which disconnects all streams.

To finish this new feature, only the UI needs to be updated.